backend/api/pyproject.toml
[project]
name = "smart-mobility-api"
version = "0.1.0"
description = "Backend FastAPI para la hackathon Smart Mobility"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.12,<1.0.0",
  "uvicorn[standard]>=0.34.0,<1.0.0",
  "sqlalchemy>=2.0.39,<3.0.0",
  "psycopg[binary]>=3.2.6,<4.0.0",
  "alembic>=1.15.1,<2.0.0",
  "firebase-admin>=6.8.0,<7.0.0",
  "pydantic-settings>=2.8.1,<3.0.0",
  "httpx>=0.28.1,<1.0.0",
  "networkx>=3.4.2,<4.0.0",
  "google-auth>=2.39.0,<3.0.0"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]